Place GUI-related files in new folder
