+++ src/main/java/Launcher.java
@@ -1,8 +1,9 @@
-import duke.Duke;
+import duke.ui.Gui;
+
 import javafx.application.Application;
 
 public class Launcher {
     public static void main(String[] args){
-        Application.launch(Duke.class, args);
+        Application.launch(Gui.class, args);
     }
 }
